Move creation of pairing partner from user input to separate class
